// src/client/src/ui/RecommendedProducts/Carousel/models.ts
//    Copyright 2020 EPAM Systems, Inc.
//
//    Licensed under the Apache License, Version 2.0 (the "License");
//    you may not use this file except in compliance with the License.
//    You may obtain a copy of the License at
//
//      http://www.apache.org/licenses/LICENSE-2.0
//
//    Unless required by applicable law or agreed to in writing, software
//    distributed under the License is distributed on an "AS IS" BASIS,
//    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
//    See the License for the specific language governing permissions and
//    limitations under the License.

import Swiper, { SwiperOptions } from 'swiper';

export interface CarouselProps {
  options: SwiperOptions;
  buttonPreviousText?: JSX.Element | string;
  buttonNextText?: JSX.Element | string;
  className?: string;
  onSlidePrev?: (swiper: Swiper) => void;
  onSlideNext?: (swiper: Swiper) => void;
  onSlideChange?: (swiper: Swiper) => void;
}

export interface CarouselState {
  currentIndex?: number;
  prevButtonDisabled: boolean;
  nextButtonDisabled: boolean;
}